Ensure correct state when short-circuitting.
When about() or say() are called again with the current subject resp. predicate,
we want to ignore the calls, to ensure nice and compact Turtle output. But
we can only do this if we are in STATE_OBJECT. Otherwise, we can end up with
illegal transitions.

This may or may not be the cause for T133924, but it should at least help us
to get better error messages.

Bug: T133924