Skip simple strings in N3Quoter::escapeLiteral
This makes the function about 50% faster on the first 5,000 Wikidata
entities. That makes the truthy dump about 7% faster in general.

This might have more or less effect on the entire truthy dump, but
I'm not willing to benchmark this in detail.

I've also added some more tests here that pass with both the old
and the new version.

Bug: T176844
Change-Id: I7141a58a022a98373c1390ee4e336fb9ee54f6c2