Remove test for “U+110000 and above”
This test is broken under PHP 8.3. According to change I943242a471
(commit c6eb2dd04f), its addition was motivated by code coverage, rather
than any known real use cases for encoding such data; I doubt that
anything actually relies on the old behavior this test exhibited, so to
me it seems best to delete the test and consider it undefined behavior.
(Recall that U+110000 is not a real Unicode code point and is not
permitted in real UTF-8: the maximum is U+10FFFF.)

As far as I can tell, the bytes in the test case weren’t even a “valid”
encoding of any U+110000 or above character to begin with; see
T368271#10348908 for a more detailed analysis.

Bug: T368271
Change-Id: Iaa2eaa52a2c89b16b8a6fe274f347f316613581a